JAVAFIXTURE-27: Add FixtureType as main interface for specimen creation
All specimens are now based on FixtureType<T> instead of Class<T> which
allows them to be considerably simplified in that they can now create
sub-specimen on their own. Additionally the caching is now based on
FixtureType<T> instead of SpecimenType (which was solely created for
caching in the first place).
All extensions now also support generics out of the box.