JAVAFIXTURE-27: Change the main class from JavaFixture to Fixture
This new feature for generic fixture-creation turned into a complete
overhaul of the code. Its introduction simplified most of the logic and
led to the removal of almost all of the existing hacks.
Renaming JavaFixture to Fixture had two reasons:
1) Fixture sounds more concise and memorable
2) This library can be used in other JVM-based languages, like Kotlin
For compatibility reasons, the old JavaFixture-class will remain as part
of the interface, although officially deprecated.