JAVAFIXTURE-12: Introduce a new endpoint to create through constructor
This will alow the user to create a new instance by using a random
constructor, if one is available. If not, the object will be created
"empty" meaning the fields will not be filled and thus be either null
or default-value.
It is important to keep in mind, that objects created this way cannot be
customized any further.