JAVAFIXTURE-35: Ease test assertion for content of set
Since a set can only contain unique values, we cannot guarantee that
the next randomly created enum-value is not already present. For
instance if an Enum has only 3 values and a set-size of 2 is requested,
it could happen that the same value is inserted twice thus resulting
in only one element in the set.