JAVAFIXTURE-35: Add support for EnumMap
Like EnumSet, this class is a little bit of a snowflake as well in that
it also has no default constructor. While EnumSet had to be created with
all possible values through a static factory method, EnumMap can at
least be created empty, although by specifying the key-class which needs
to be an Enum.