JAVAFIXTURE-38: Extend InstanceFactory
Creating a proxy will now try to use both private constructors and
factory-methods before it tries to proxy it. This is not the ideal
solution and the usage should be refactored further. Also, this is
currently misleading because calling a method named 'proxy' that is not
always creating a proxy sounds like a lie to me. But this is a first
approach that made the test green.