JAVAFIXTURE-38: Limit factory creation to public
The fallback to find and use a factory-method if all other mechanisms
for creation failed, should be limited to public methods. If there is
no other way to create the class and even the factory-method is
protected, then this class should never be created in the first place
and JavaFixture should fail.