JAVAFIXTURE-52: Add customization by type
The builder interface now has a new method overload for .with() that
takes a type (or specimen-type) as the first argument and an instance
of the same type as a second argument. This will be added to the cache
in order to allow its injection into the resulting object tree.