JAVAFIXTURE-49: Add test to cover base class
This is related to JAVAFIXTURE-50 where fields of base-classes cannot
be set when nested. In fact, fields of any base-class cannot be set,
since they're not reflected properly right now.
The new behaviour to throw if a missing field is being customized
exposed this general problem and thus justifies a test.