refactor: check for valid candidate before picking
Since we pick a random class from the found implementations, we want to
filter out unsuitable candidates before we pick one. Otherwise, we might
pick one that can't be used while we skip other candidates which would
have worked.
This also resolves a potential NPE when creating a GenericSpecimen with
mismatching type-parameters.