replace random by strategic specimen-selection
Instead of picking a random implementation candidate for an abstract
type at build-time of the specimen, all matching candidates are passed
to the specimen and iterated over at creation-time until a suitable
implementation is found. This allows to try multiple implementations and
to fall back to a proxy if none of the candidates can be manufactured.