fix: pick constructor matching customization
When a field is customized, using a random constructor might pick one
that does not contain the customized field therefore ignoring the
customization. Instead, we select all constructors that contain the
customized fields and pick a random one out of those. If no constructor
with the customized fields is found, an exception is thrown. This will
only work for records since traditional classes do not store proper
constructor argument names (instead use arg0, arg1, etc.) and cannot
match those to customized field-names.