fix: improve error-message for unknown nested field
When a nested field is customized (e.g. `.with("parent.xxx", 42)`), but
the field `xxx` doesn't exist in `parent`, a proper exception is thrown,
but then misinterpreted as a failure to construct the object via
reflection, then retried using the constructor which then fails
ultimately with a misleading message. Instead, we introduced a
specialized exception for customization-validation that will bubble up
with the proper error-message without being misinterpreted.